+++ 2019-Robot/src/main/java/frc/team670/robot/utils/math/AngleStorage.java
@@ -1,0 +1,18 @@
+package frc.team670.robot.utils.math;
+
+
+public class AngleStorage {
+		private double angle;
+
+		public AngleStorage(double angle) {
+			setAngle(angle);
+		}
+
+		public void setAngle(double angle) {
+			this.angle = angle;
+		}
+
+		public double getAngle() {
+			return angle;
+		}
+}
